Updated Event Manager
Similar to the WindowManager Update, the EventManager is also solely owned and is a property of the GameManager. Therefore, the Singleton Design Pattern has been removed from EventManager. However, just like for the WindowManager, the EventManager can still be accessed globally through the GameManager because it still follows the Singleton Design Pattern and can therefore be accessed globally.
- Updated EventManager to handle both Main Window and Debug Window Events based on the situation. For example, if the Debug Window is open, don't process Main Window Events (except for the Close Event).
- GameManager now owns EventManager and fully handles the game logic within the process() function of GameManager.