Added centerCameraView() Function
The new centerCameraView() function puts centers the camera relative to the position of the Game Board. This function is called BEFORE the Game Loop to ensure that the Camera View starts at the center of the board. This is important for a future Camera View clamping feature which will clamp Camera View movements to the Game Board size and position (accounting for the margin).
- Added centerCameraView() function.
- Grid now always positions itself based on the Margin within its constructor.
- Removed Grid positioning logic in WindowManager.draw() function.