Added Base Resolutions
Added base resolutions for both 4:3 and 16:9 aspect ratios. This allows for consistency in the zooming to allow for the user to resize their screen assuming that they stay at either 4:3 or 16:9 aspect ratios.
- Added base 4:3 resolution.
- Added base 16:9 resolution.

I plan to add a settings menu which will allow the user to resize their screen based on a list of resolutions that either have 4:3 or 16:9 aspect ratios. This will ensure consistency in the aspect ratios and therefore keep visuals consistent.