TileMap System
Created the a TileMap system which handles loading a Tile Set into a Texture from a path. The TileMap system is responsible for section the loaded Tile Set Texture into different little Tiles based on the size of each Tiles. This would be memory efficient since there is only a singular loaded Tile Set.

Properties:
- Added m_tileMapTexture.
- Added m_errorTexture.
- Added m_tileSize.
- Added m_tileMapSize.
- Added m_tileMapSizeInTiles.
- Added m_totalTiles.
- Added m_hasLoadedTexture.

Functions:
- Added TileMap Constructors.
- Added TileMap Destructor.
- Added loadTileMap().
- Added getTile().
- Added getTileMapTexture().
- Added getTileSize().
- Added getTileMapSize().
- Added getTileMapSizeInTiles().
- Added getTotalTiles().
- Added assertXY().